% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{add_issues_for_suso_errors}
\alias{add_issues_for_suso_errors}
\title{Add an issue for each SuSo validation error}
\usage{
add_issues_for_suso_errors(
  df_cases_to_review,
  df_errors,
  issue_type = 3,
  df_issues
)
}
\arguments{
\item{df_cases_to_review}{Data frame containing cases to review.}

\item{df_errors}{Data frame containing SuSo validation errors.}

\item{issue_type}{Numeric. Value of the issue type that each SuSo error will be considered.}

\item{df_issues}{Data frame containing auto-reject issues.}
}
\value{
Data frame. Issues file with additional rows for SuSo validation errors.
}
\description{
First, transform SuSo validations from \code{interview__errors} into issues of the type indicated in \code{issue_type}.
Then add these issues to the main issues files.
}
