% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{add_rejection_msgs}
\alias{add_rejection_msgs}
\title{Add consolidated rejection message to data frame of interviews to reject}
\usage{
add_rejection_msgs(df_to_reject, df_issues, issue_types_to_reject = 1)
}
\arguments{
\item{df_to_reject}{Data frame. Interviews to reject. Contains columns interview__id, interview__key, interviewComplete, interview__status}

\item{df_issues}{Data frame. Issues noted for each interview.}

\item{issue_types_to_reject}{Numeric. Issue type values to reject.}
}
\value{
Data frame with 1 observation per interview that consists of interview ID, rejection message, and interview status
}
\description{
Add consolidated rejection message to data frame of interviews to reject
}
