% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{any_obs}
\alias{any_obs}
\title{Determine whether where a condition is \code{TRUE} for any observation.}
\usage{
any_obs(df, where, attrib_name, attrib_vars)
}
\arguments{
\item{df}{Data frame. Household-level microdata that contains the attribute to extract.}

\item{where}{Logical expression whose variables exist in \code{df}.}

\item{attrib_name}{Character. Name to give the attribute in the issues data file}

\item{attrib_vars}{Character. Regular expression that identifies the variable in \code{var}.}
}
\value{
Data frame of attributes with the following columns:
\itemize{
\item \code{interview__id}
\item \code{interview__key}
\item \code{attrib_name}
\item \code{attrib_val}
\item \code{attrib_vars}
}
}
\description{
Determine whether a user-provided condition is \code{TRUE} for any non-household
level microdata--such as a parcel, plot, of food item data frame.
}
