% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{any_vars}
\alias{any_vars}
\title{Create attribute about whether any variable contains a specific value}
\usage{
any_vars(df, var_pattern, var_val = 1, attrib_name, attrib_vars = var_pattern)
}
\arguments{
\item{df}{Data frame. Household-level microdata that contains the attribute to extract.}

\item{var_pattern}{Character Regular expression used to select variables to be counted.}

\item{var_val}{Numeric. Value to count in columns identified by \code{var_pattern}.}

\item{attrib_name}{Character. Name to give the attribute in the issues data file}

\item{attrib_vars}{Character. Regular expression that identifies the variable in \code{var}.}
}
\value{
Data frame of attributes with the following columns:
\itemize{
\item \code{interview__id}
\item \code{interview__key}
\item \code{attrib_name}
\item \code{attrib_val}
\item \code{attrib_vars}
}
}
\description{
Inspect all columns whose names match \code{var_pattern}
and determine whether any contain the value \code{var_val}.
}
