% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{approve_interview}
\alias{approve_interview}
\title{Approve interviews using the appropriate set of API calls}
\usage{
approve_interview(
  interview__id,
  interview__status,
  approve_comment,
  statuses_to_approve = c(120),
  server = Sys.getenv("SUSO_SERVER"),
  workspace = Sys.getenv("SUSO_WORKSPACE"),
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)
}
\arguments{
\item{interview__id}{Character. GUID for interview found in \code{interview__id}.}

\item{interview__status}{Numeric. Supports values in set c(100, 120).}

\item{approve_comment}{Character. Comment to post upon rejection}

\item{statuses_to_approve}{Numeric vector. Supports values in set c(100, 120).}

\item{server}{Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}

\item{workspace}{Character. Name of the workspace whose questionnaires and associated interviews to get.}

\item{user}{API user name}

\item{password}{API password}
}
\description{
Approve interviews using the appropriate set of API calls
}
