% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{count_list}
\alias{count_list}
\title{Count the non-missing elements of a list variable}
\usage{
count_list(
  df,
  var_pattern,
  missing_vals = c("##N/A##", "", NA_character_),
  attrib_name,
  attrib_vars = var_pattern
)
}
\arguments{
\item{df}{Data frame. Household-level microdata that contains the attribute to extract.}

\item{var_pattern}{Character. Regular expression used to select variables to be counted.}

\item{missing_vals}{Character. Values that describes an empty element of a list variable.}

\item{attrib_name}{Character. Name to give the attribute in the issues data file}

\item{attrib_vars}{Character. Regular expression that identifies the variable in \code{var}.}
}
\value{
Data frame of attributes with the following columns:
\itemize{
\item \code{interview__id}
\item \code{interview__key}
\item \code{attrib_name}
\item \code{attrib_val}
\item \code{attrib_vars}
}
}
\description{
Count the non-missing elements of a list variable
}
