% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{count_vars}
\alias{count_vars}
\title{Count columns with value in microdata}
\usage{
count_vars(
  df,
  var_pattern,
  var_val = 1,
  attrib_name,
  attrib_vars = var_pattern
)
}
\arguments{
\item{df}{Data frame. Household-level microdata that contains the variables to count.}

\item{var_pattern}{Character Regular expression used to select variables to be counted.}

\item{var_val}{Numeric. Value(s) to count in columns identified by \code{var_pattern}.}

\item{attrib_name}{Character. Name of attribute.}

\item{attrib_vars}{Character. Regular expression that identifies the variable(s) in \code{var_pattern}.}
}
\value{
Data frame of attributes with the following columns:
\itemize{
\item \code{interview__id}
\item \code{interview__key}
\item \code{attrib_name}
\item \code{attrib_val}
\item \code{attrib_vars}
}
}
\description{
Count then number of columns matching \code{var_pattern} that have value \code{var_val}.
}
