% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{create_issue}
\alias{create_issue}
\title{Create an issue from one or more attributes in attribute data}
\usage{
create_issue(df_attribs, vars, where, type = 1, desc, comment)
}
\arguments{
\item{df_attribs}{Data frame of attributes}

\item{vars}{Character. Variable(s) needed for the logical expression in \code{condition}.}

\item{where}{Logical expression that if \code{TRUE} indicates an issue.}

\item{type}{Type of issue. Values are as follows: c(\code{Reject} = 1, \code{Comment} = 2, \code{Review} = 4)}

\item{desc}{Character. Short description for survey managers to understand the issue.}

\item{comment}{Character. Longer message for interviewers to understand the issue and what to do about it.}
}
\value{
Data frame of issues with the following columns:
\itemize{
\item \code{interview__id}
\item \code{interview__key}
\item \code{issue_type}
\item \code{issue_desc}
\item \code{issue_comment}
\item \code{issue_vars}
\item \code{issue_loc}
}
}
\description{
Create an issue from one or more attributes in attribute data
}
