% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{extract_attribute}
\alias{extract_attribute}
\title{Extract attributes from variables in microdata}
\usage{
extract_attribute(df, var, attrib_name, attrib_vars)
}
\arguments{
\item{df}{Data frame. Household-level microdata that contains the attribute to extract.}

\item{var}{Bare name. Variable whose value to extract.}

\item{attrib_name}{Character. Name to give the attribute in the issues data file}

\item{attrib_vars}{Character. Regular expression that identifies the variable in \code{var}.}
}
\value{
Data frame of attributes with the following columns:
\itemize{
\item \code{interview__id}
\item \code{interview__key}
\item \code{attrib_name}
\item \code{attrib_val}
\item \code{attrib_vars}
}
}
\description{
Extract attributes from variables in microdata
}
