% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{flag_persistent_issues}
\alias{flag_persistent_issues}
\title{Flag persistant issues}
\usage{
flag_persistent_issues(df_comments, df_to_reject)
}
\arguments{
\item{df_comments}{Data frame. Comments on interviews. Importantly, this includes comments posted upon rejection.}

\item{df_to_reject}{Data frame. Interviews to be rejected.}
}
\value{
List of data frames: list(to_reject, to_follow_up). The to_reject data frame removes any cases where follow-up is required for persistent problems. The to_follow_up data frame identifies any cases where follow-up is required for persistent problems.
}
\description{
Use rejection comments to flag persistent issues--that is, rejection reasons that been repeated for the interview.
Creates a data frame of interviews with persistent issues. These require manual follow-up by headquarters staff.
Removes any interviews with persistent isues from the to-reject data frame.
}
