% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{make_issue}
\alias{make_issue}
\title{Make an issue from values in household-level microdata}
\usage{
make_issue(df, where, type = 1, desc, comment, issue_vars)
}
\arguments{
\item{df}{Data frame of microdata}

\item{where}{Logical expression that if \code{TRUE} indicates an issue.}

\item{type}{Type of issue. Values are as follows: c(\code{Reject} = 1, \code{Comment} = 2, \code{Review} = 4)}

\item{desc}{Character. Short description for survey managers to understand the issue.}

\item{comment}{Character. Longer message for interviewers to understand the issue and what to do about it.}

\item{issue_vars}{Character. Regular expression to describe variables involved in issue.}
}
\value{
Data frame of issues with the following columns:
\itemize{
\item \code{interview__id}
\item \code{interview__key}
\item \code{issue_type}
\item \code{issue_desc}
\item \code{issue_comment}
\item \code{issue_vars}
\item \code{issue_loc}
}
}
\description{
Make an issue from values in household-level microdata
}
