% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{post_comments}
\alias{post_comments}
\title{Post comments to questions, if any}
\usage{
post_comments(
  df_to_reject,
  df_issues,
  server = Sys.getenv("SUSO_SERVER"),
  workspace = "primary",
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)
}
\arguments{
\item{df_to_reject}{Data frame containing interviews to reject.}

\item{df_issues}{Data frame of issues.}

\item{server}{Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}

\item{workspace}{Character. Name of the workspace whose questionnaires and associated interviews to get.}

\item{user}{API user name}

\item{password}{API password}
}
\value{
Server-side effect of posting all applicable comments to all applicable questionnaires
}
\description{
First, extracts comments from issues file for interviews to reject. Then, posts those comments, if any
}
