% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{reject_interview}
\alias{reject_interview}
\title{Reject interview using the appropriate set of rejection API calls}
\usage{
reject_interview(
  interview__id,
  interview__status,
  reject_comment,
  statuses_to_reject = c(100, 120),
  reject_hq_approved = TRUE,
  server = Sys.getenv("SUSO_SERVER"),
  workspace = "primary",
  user = Sys.getenv("SUSO_USER"),
  password = Sys.getenv("SUSO_PASSWORD")
)
}
\arguments{
\item{interview__id}{Character. GUID for interview found in \code{interview__id}.}

\item{interview__status}{Numeric. Supports values in set c(100, 120, 130).}

\item{reject_comment}{Character. Comment to post upon rejection}

\item{statuses_to_reject}{Numeric vector. Supports values in set c(100, 120, 130).}

\item{reject_hq_approved}{Boolean. Flag that determines whether HeadquartersApproved assignments will be rejected or not. Default behavior is to reject.}

\item{server}{Full server web address (e.g., \code{https://demo.mysurvey.solutions}, \code{https://my.domain})}

\item{workspace}{Character. Name of the workspace whose questionnaires and associated interviews to get.}

\item{user}{API user name}

\item{password}{API password}
}
\description{
Reject interview using the appropriate set of rejection API calls
}
